add cancel button and refactor adapter
