+++ app/src/main/java/org/wit/archaeologicalfieldwork/activities/HillfortAdapter.kt
@@ -1,0 +1,34 @@
+package org.wit.archaeologicalfieldwork.activities
+
+import android.support.v7.widget.RecyclerView
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import kotlinx.android.synthetic.main.card_hillfort.view.*
+import org.jetbrains.anko.AnkoLogger
+import org.jetbrains.anko.info
+import org.wit.archaeologicalfieldwork.R
+import org.wit.archaeologicalfieldwork.models.HillfortModel
+
+class HillfortAdapter constructor(private var hillforts: List<HillfortModel>) : RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
+
+  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
+    return MainHolder(LayoutInflater.from(parent?.context).inflate(R.layout.card_hillfort, parent, false))
+  }
+
+  override fun onBindViewHolder(holder: MainHolder, position: Int) {
+    val hillfort = hillforts[holder.adapterPosition]
+    holder.bind(hillfort)
+  }
+
+  override fun getItemCount(): Int = hillforts.size
+
+  class MainHolder constructor(itemView: View) : RecyclerView.ViewHolder(itemView), AnkoLogger {
+
+    fun bind(hillfort: HillfortModel) {
+      info(hillfort.name)
+      itemView.hillfortName.text = hillfort.name
+      itemView.description.text = hillfort.description
+    }
+  }
+}
