+++ app/src/main/java/org/wit/archaeologicalfieldwork/activities/HillfortListActivity.kt
@@ -12,8 +12,7 @@ import org.jetbrains.anko.info
 import org.jetbrains.anko.startActivityForResult
 import org.wit.archaeologicalfieldwork.R
 import org.wit.archaeologicalfieldwork.main.MainApp
-import org.wit.archaeologicalfieldwork.models.HillfortModel
-import kotlin.math.log
+
 
 class HillfortListActivity : AppCompatActivity() {
 
@@ -44,25 +43,3 @@ class HillfortListActivity : AppCompatActivity() {
   }
 }
 
-class HillfortAdapter constructor(private var hillforts: List<HillfortModel>) : RecyclerView.Adapter<HillfortAdapter.MainHolder>() {
-
-  override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MainHolder {
-    return MainHolder(LayoutInflater.from(parent?.context).inflate(R.layout.card_hillfort, parent, false))
-  }
-
-  override fun onBindViewHolder(holder: MainHolder, position: Int) {
-    val hillfort = hillforts[holder.adapterPosition]
-    holder.bind(hillfort)
-  }
-
-  override fun getItemCount(): Int = hillforts.size
-
-  class MainHolder constructor(itemView: View) : RecyclerView.ViewHolder(itemView), AnkoLogger {
-
-    fun bind(hillfort: HillfortModel) {
-      info(hillfort.name)
-      itemView.hillfortName.text = hillfort.name
-      itemView.description.text = hillfort.description
-    }
-  }
-}
