+++ app/src/main/res/menu/menu_hillfort.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:app="http://schemas.android.com/apk/res-auto">
+
+  <item
+    android:id="@+id/item_cancel"
+    android:title="@string/menu_cancelHillfort"
+    app:showAsAction="always"/>
+</menu>
