+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="hint_HillFortTitle">Hillfort Title</string>
     <string name="hint_HillFortDescription">Description</string>
     <string name="menu_addHillfort">Add</string>
+    <string name="menu_cancelHillfort">Cancel</string>
 </resources>
